refactor: HttpResponse 내부에 forwardByTemplate 구현
