+++ src/main/java/http/model/Cookie.java
@@ -7,17 +7,17 @@ public class Cookie {
     Map<String, String> cookieMap;
 
     public Cookie(String cookies) {
+        cookieMap = new HashMap<>();
+
         if (cookies == null) {
             return;
         }
         String[] eachCookies = cookies.split("; ");
 
-        Map<String, String> cookieMap = new HashMap<>();
         for (String eachCookie : eachCookies) {
             String[] cookiePair = eachCookie.split("=");
             cookieMap.put(cookiePair[0], cookiePair[1]);
         }
-        this.cookieMap = cookieMap;
     }
 
     public boolean hasCookie(String key) {
