+++ src/main/java/http/model/HttpResponse.java
@@ -1,6 +1,13 @@
 package http.model;
 
+import com.github.jknack.handlebars.Handlebars;
+import com.github.jknack.handlebars.Template;
+import com.github.jknack.handlebars.io.ClassPathTemplateLoader;
+import com.github.jknack.handlebars.io.TemplateLoader;
 import utils.FileIoUtils;
+
+import java.io.IOException;
+import java.util.Map;
 
 import static com.google.common.net.HttpHeaders.*;
 
@@ -83,16 +90,40 @@ public class HttpResponse {
             return this;
         }
 
-        public Builder sendRedirect(String filePath) {
-            httpHeaders.addHeader(LOCATION, ROOT_URI + filePath);
+        public Builder sendRedirect(String url) {
+            httpHeaders.addHeader(LOCATION, ROOT_URI + url);
             protocols(HttpProtocols.HTTP1_1);
             status(HttpStatus.FOUND);
             addHeader(CONTENT_TYPE, ContentType.HTML.getType());
             return this;
         }
 
+        public Builder forwardByTemplate(String url, Map<String, Object> model) {
+            try {
+                Handlebars handlebars = getHandlebars();
+                Template template = handlebars.compile(url);
+                String page = template.apply(model);
+                this.body = page.getBytes();
+                protocols(HttpProtocols.HTTP1_1);
+                status(HttpStatus.OK);
+                addHeader(CONTENT_TYPE, ContentType.HTML.getType());
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+            return this;
+        }
+
         public HttpResponse build() {
             return new HttpResponse(this);
+        }
+
+        private Handlebars getHandlebars() {
+            TemplateLoader loader = new ClassPathTemplateLoader();
+            loader.setPrefix("/templates");
+            loader.setSuffix(".html");
+            Handlebars handlebars = new Handlebars(loader);
+            handlebars.registerHelper("idx", ((context, options) -> (Integer) context + 1));
+            return handlebars;
         }
     }
 }
